requirements.txt
requests>=2.28.0
psutil>=5.9.0
pywin32>=305
WMI>=1.5.1
pycaw>=20230407
comtypes>=1.2.0
winrt-Windows.Media.Control>=2.0.0
winrt-Windows.Foundation>=2.0.0
python-dotenv>=1.0.0
pystray>=0.19.5
Pillow>=10.0.0
